requirements.txt
PyGithub==1.26.0
requests==2.10.0
